// src/stores/chat/interface.ts
export interface ChatState {
    isPopupDl: boolean;
    isPopupPay: boolean;
    isPopupCoin: boolean;
    isPopupAi: boolean;
    aiInfo: any;
    aiList: any;
    chatIm: any;
    sdkAppId: number;
    init: any;
    chatList: any;
    isSend: number;
    originalData: any;
    isScroll: boolean;
    isAI: boolean;
    isLoadAi: boolean;
    payToken: any;
    delAiInfo: any;
}